+++ test/e2e/atlas/backup_export_buckets_test.go
@@ -1,0 +1,121 @@
+// Copyright 2022 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//go:build e2e || (atlas && backup)
+
+package atlas_test
+
+import (
+	"encoding/json"
+	"os"
+	"os/exec"
+	"testing"
+
+	"github.com/mongodb/mongodb-atlas-cli/test/e2e"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	atlas "go.mongodb.org/atlas/mongodbatlas"
+)
+
+func TestExportBuckets(t *testing.T) {
+	cliPath, err := e2e.AtlasCLIBin()
+	r := require.New(t)
+	r.NoError(err)
+
+	const cloudProvider = "AWS"
+	iamRoleID := os.Getenv("E2E_CLOUD_ROLE_ID")
+	bucketName := os.Getenv("E2E_TEST_BUCKET")
+	r.NotEmpty(iamRoleID)
+	r.NotEmpty(bucketName)
+	var bucketID string
+
+	t.Run("Create", func(t *testing.T) {
+		cmd := exec.Command(cliPath,
+			backupsEntity,
+			exportsEntity,
+			bucketsEntity,
+			"create",
+			bucketName,
+			"--cloudProvider",
+			cloudProvider,
+			"--iamRoleId",
+			iamRoleID,
+			"-o=json")
+		cmd.Env = os.Environ()
+		resp, err := cmd.CombinedOutput()
+
+		r.NoError(err, string(resp))
+
+		a := assert.New(t)
+		var exportBucket atlas.CloudProviderSnapshotExportBucket
+		if err = json.Unmarshal(resp, &exportBucket); a.NoError(err) {
+			a.Equal(bucketName, exportBucket.BucketName)
+		}
+		bucketID = exportBucket.ID
+	})
+
+	t.Run("List", func(t *testing.T) {
+		cmd := exec.Command(cliPath,
+			backupsEntity,
+			exportsEntity,
+			bucketsEntity,
+			"list",
+			"-o=json")
+		cmd.Env = os.Environ()
+		resp, err := cmd.CombinedOutput()
+
+		r.NoError(err, string(resp))
+
+		var r atlas.CloudProviderSnapshotExportBuckets
+		a := assert.New(t)
+		if err = json.Unmarshal(resp, &r); a.NoError(err) {
+			a.NotEmpty(r)
+		}
+	})
+
+	t.Run("Describe", func(t *testing.T) {
+		cmd := exec.Command(cliPath,
+			backupsEntity,
+			exportsEntity,
+			bucketsEntity,
+			"describe",
+			"--bucketId",
+			bucketID,
+			"-o=json")
+		cmd.Env = os.Environ()
+		resp, err := cmd.CombinedOutput()
+
+		r.NoError(err, string(resp))
+
+		a := assert.New(t)
+		var exportBucket atlas.CloudProviderSnapshotExportBucket
+		if err = json.Unmarshal(resp, &exportBucket); a.NoError(err) {
+			a.Equal(bucketName, exportBucket.BucketName)
+		}
+	})
+
+	t.Run("Delete", func(t *testing.T) {
+		cmd := exec.Command(cliPath,
+			backupsEntity,
+			exportsEntity,
+			bucketsEntity,
+			"delete",
+			"--bucketId",
+			bucketID,
+			"--force")
+		cmd.Env = os.Environ()
+		resp, err := cmd.CombinedOutput()
+
+		r.NoError(err, string(resp))
+	})
+}
