+++ test/e2e/atlas/backup_export_jobs_test.go
@@ -28,99 +28,6 @@ import (
 	atlas "go.mongodb.org/atlas/mongodbatlas"
 )
 
-func TestExportBuckets(t *testing.T) {
-	cliPath, err := e2e.AtlasCLIBin()
-	r := require.New(t)
-	r.NoError(err)
-
-	const cloudProvider = "AWS"
-	iamRoleID := os.Getenv("E2E_CLOUD_ROLE_ID")
-	bucketName := os.Getenv("E2E_TEST_BUCKET")
-	r.NotEmpty(iamRoleID)
-	r.NotEmpty(bucketName)
-	var bucketID string
-
-	t.Run("Create", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			exportsEntity,
-			bucketsEntity,
-			"create",
-			bucketName,
-			"--cloudProvider",
-			cloudProvider,
-			"--iamRoleId",
-			iamRoleID,
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-
-		r.NoError(err, string(resp))
-
-		a := assert.New(t)
-		var exportBucket atlas.CloudProviderSnapshotExportBucket
-		if err = json.Unmarshal(resp, &exportBucket); a.NoError(err) {
-			a.Equal(bucketName, exportBucket.BucketName)
-		}
-		bucketID = exportBucket.ID
-	})
-
-	t.Run("List", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			exportsEntity,
-			bucketsEntity,
-			"list",
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-
-		r.NoError(err, string(resp))
-
-		var r atlas.CloudProviderSnapshotExportBuckets
-		a := assert.New(t)
-		if err = json.Unmarshal(resp, &r); a.NoError(err) {
-			a.NotEmpty(r)
-		}
-	})
-
-	t.Run("Describe", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			exportsEntity,
-			bucketsEntity,
-			"describe",
-			"--bucketId",
-			bucketID,
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-
-		r.NoError(err, string(resp))
-
-		a := assert.New(t)
-		var exportBucket atlas.CloudProviderSnapshotExportBucket
-		if err = json.Unmarshal(resp, &exportBucket); a.NoError(err) {
-			a.Equal(bucketName, exportBucket.BucketName)
-		}
-	})
-
-	t.Run("Delete", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			exportsEntity,
-			bucketsEntity,
-			"delete",
-			"--bucketId",
-			bucketID,
-			"--force")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-
-		r.NoError(err, string(resp))
-	})
-}
-
 func TestExportJobs(t *testing.T) {
 	cliPath, err := e2e.AtlasCLIBin()
 	r := require.New(t)
@@ -338,167 +245,6 @@ func TestExportJobs(t *testing.T) {
 	})
 
 	t.Run("watch snapshot deletion", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			snapshotsEntity,
-			"watch",
-			snapshotID,
-			"--clusterName",
-			clusterName)
-		cmd.Env = os.Environ()
-		resp, _ := cmd.CombinedOutput()
-		t.Log(string(resp))
-	})
-
-	t.Run("Delete cluster", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			clustersEntity,
-			"delete",
-			clusterName,
-			"--force",
-		)
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-		r.NoError(err, string(resp))
-
-		expected := fmt.Sprintf("Cluster '%s' deleted\n", clusterName)
-		a := assert.New(t)
-		a.Equal(expected, string(resp))
-	})
-
-	t.Run("Watch delete cluster", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			clustersEntity,
-			"watch",
-			clusterName,
-		)
-		cmd.Env = os.Environ()
-		resp, _ := cmd.CombinedOutput()
-		t.Log(string(resp))
-	})
-}
-
-func TestSnapshots(t *testing.T) {
-	cliPath, err := e2e.AtlasCLIBin()
-	r := require.New(t)
-	r.NoError(err)
-
-	clusterName, err := RandClusterName()
-	r.NoError(err)
-	fmt.Println(clusterName)
-
-	var snapshotID string
-
-	t.Run("Create cluster", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			clustersEntity,
-			"create",
-			clusterName,
-			"--backup",
-			"--tier", tierM10,
-			"--region=US_EAST_1",
-			"--provider", e2eClusterProvider,
-			"--mdbVersion", e2eSharedMDBVer,
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-		r.NoError(err, string(resp))
-
-		var cluster *atlas.Cluster
-		err = json.Unmarshal(resp, &cluster)
-		r.NoError(err)
-
-		ensureSharedCluster(t, cluster, clusterName, e2eSharedMDBVer, tierM10, 10, false)
-	})
-
-	t.Run("Watch create cluster", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			clustersEntity,
-			"watch",
-			clusterName,
-		)
-		cmd.Env = os.Environ()
-		resp, _ := cmd.CombinedOutput()
-		t.Log(string(resp))
-	})
-
-	t.Run("Create snapshot", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			snapshotsEntity,
-			"create",
-			clusterName,
-			"--desc",
-			"test-snapshot",
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-
-		r.NoError(err, string(resp))
-
-		a := assert.New(t)
-		var snapshot atlas.CloudProviderSnapshot
-		if err = json.Unmarshal(resp, &snapshot); a.NoError(err) {
-			a.Equal("test-snapshot", snapshot.Description)
-		}
-		snapshotID = snapshot.ID
-	})
-
-	t.Run("List", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			snapshotsEntity,
-			"list",
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-
-		r.NoError(err, string(resp))
-
-		var r atlas.CloudProviderSnapshots
-		a := assert.New(t)
-		if err = json.Unmarshal(resp, &r); a.NoError(err) {
-			a.NotEmpty(r)
-		}
-	})
-
-	t.Run("Describe", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			snapshotsEntity,
-			"list",
-			snapshotID,
-			"--clusterName",
-			clusterName,
-			"-o=json")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-
-		r.NoError(err, string(resp))
-
-		a := assert.New(t)
-		var result atlas.CloudProviderSnapshot
-		if err = json.Unmarshal(resp, &result); a.NoError(err) {
-			a.Equal(snapshotID, result.ID)
-		}
-	})
-
-	t.Run("Delete", func(t *testing.T) {
-		cmd := exec.Command(cliPath,
-			backupsEntity,
-			snapshotsEntity,
-			"delete",
-			snapshotID,
-			"--clusterName",
-			clusterName,
-			"--force")
-		cmd.Env = os.Environ()
-		resp, err := cmd.CombinedOutput()
-
-		r.NoError(err, string(resp))
-	})
-
-	t.Run("Watch deletion", func(t *testing.T) {
 		cmd := exec.Command(cliPath,
 			backupsEntity,
 			snapshotsEntity,
